+++ knishioclient/mutation/MutationWithdrawBufferToken.py
@@ -7,13 +7,12 @@ class MutationWithdrawBufferToken(MutationProposeMolecule):
     Mutation for withdrawing tokens from buffer
     """
     
-    def fill_molecule(self, recipients: dict, signing_wallet=None):
+    def fill_molecule(self, recipients: dict):
         """
         Fills the Molecule with withdrawal data
         
         :param recipients: Dict of recipient_bundle: amount mappings
-        :param signing_wallet: Optional signing wallet
         """
-        self.molecule().init_withdraw_buffer(recipients, signing_wallet)
+        self.molecule().init_withdraw_buffer(recipients)
         self.molecule().sign()
         self.molecule().check()